starting over with new main.rs
