+++ Cargo.toml
@@ -16,7 +16,7 @@ wasm-bindgen = "^0.2"
 web-sys = { version = "^0.3.60", features = [ "Element", "MouseEvent", "DomRect" ] }
 yew = "^0.19"
 gloo = "^0.8"
-gloo-console = "^0.2"
+leptos = { version = "0.6.12", features = ["csr", "nightly"] }
 
 [profile.release]
 lto = true
